Refactored some profile edition and profile retrieving views
